+++ cmd/workflow-engine/cli/pipelines.go
@@ -56,16 +56,14 @@ func newRunCommand() *cobra.Command {
 	imageScanCmd.Flags().String("grype-filename", "", "the output filename for the grype vulnerability report")
 	_ = viper.BindPFlag("imagescan.grypefilename", imageScanCmd.Flags().Lookup("grype-filename"))
 
+	imageScanCmd.Flags().String("clamav-filename", "", "the output filename for the ClamAV scan report")
+	_ = viper.BindPFlag("imagescan.clamavfilename", imageScanCmd.Flags().Lookup("clamav-filename"))
+
 	imageScanCmd.Flags().String("scan-image-target", "", "scan a specific image")
 	_ = viper.BindPFlag("imagescan.targetimage", imageScanCmd.Flags().Lookup("scan-image-target"))
 
-	imageScanCmd.Flags().String("clamav-filename", "", "the output filename for the ClamAV scan report")
-	_ = viper.BindPFlag("artifacts.clamavfilename", imageScanCmd.Flags().Lookup("clamav-filename"))
-	viper.MustBindEnv("artifacts.clamavfilename", "WFE_CLAMAV_FILENAME")
-
 	// run image-publish
 	imagePublishCmd := newBasicCommand("image-publish", "publishes an image", runimagePublish)
-	imagePublishCmd.Flags().Bool("no-push", false, "disable pushing the images to the repository, intended for local developement")
 
 	// run code-scan
 	codeScanCmd := newBasicCommand("code-scan", "run Static Application Security Tests (SAST) scans", runCodeScan)
@@ -80,6 +78,10 @@ func newRunCommand() *cobra.Command {
 	_ = viper.BindPFlag("codescan.semgreprules", codeScanCmd.Flags().Lookup("semgrep-rules"))
 
 	codeScanCmd.Flags().Bool("semgrep-experimental", false, "use the osemgrep statically compiled binary")
+
+	// run deploy
+	deployCmd := newBasicCommand("deploy", "Beta Feature: VALIDATION ONLY - run gatecheck validate on artifacts from previous pipelines", runDeploy)
+
 	// run
 	cmd := &cobra.Command{Use: "run", Short: "run a pipeline"}
 
@@ -96,7 +98,7 @@ func newRunCommand() *cobra.Command {
 	cmd.SilenceUsage = true
 
 	// Add sub commands
-	cmd.AddCommand(debugCmd, imageBuildCmd, imageScanCmd, imagePublishCmd, codeScanCmd)
+	cmd.AddCommand(debugCmd, imageBuildCmd, imageScanCmd, imagePublishCmd, codeScanCmd, deployCmd)
 
 	return cmd
 }
@@ -106,6 +108,18 @@ func newRunCommand() *cobra.Command {
 func runDebug(cmd *cobra.Command, _ []string) error {
 	dryRunEnabled, _ := cmd.Flags().GetBool("dry-run")
 	return debugPipeline(cmd.OutOrStdout(), cmd.ErrOrStderr(), dryRunEnabled)
+}
+
+func runDeploy(cmd *cobra.Command, _ []string) error {
+	dryRunEnabled, _ := cmd.Flags().GetBool("dry-run")
+	configFilename, _ := cmd.Flags().GetString("config")
+
+	config := new(pipelines.Config)
+	if err := LoadOrDefault(configFilename, config, viper.GetViper()); err != nil {
+		return err
+	}
+
+	return deployPipeline(cmd.OutOrStdout(), cmd.ErrOrStderr(), config, dryRunEnabled)
 }
 
 func runImageBuild(cmd *cobra.Command, _ []string) error {
@@ -136,13 +150,12 @@ func runImageScan(cmd *cobra.Command, _ []string) error {
 func runimagePublish(cmd *cobra.Command, _ []string) error {
 	dryRunEnabled, _ := cmd.Flags().GetBool("dry-run")
 	configFilename, _ := cmd.Flags().GetString("config")
-	nopush, _ := cmd.Flags().GetBool("no-push")
 
 	config := new(pipelines.Config)
 	if err := LoadOrDefault(configFilename, config, viper.GetViper()); err != nil {
 		return err
 	}
-	return imagePublishPipeline(cmd.OutOrStdout(), cmd.ErrOrStderr(), config, dryRunEnabled, nopush)
+	return imagePublishPipeline(cmd.OutOrStdout(), cmd.ErrOrStderr(), config, dryRunEnabled)
 }
 
 func runCodeScan(cmd *cobra.Command, _ []string) error {
@@ -176,10 +189,9 @@ func imageScanPipeline(stdout io.Writer, stderr io.Writer, config *pipelines.Con
 	return pipeline.WithConfig(config).Run()
 }
 
-func imagePublishPipeline(stdout io.Writer, stderr io.Writer, config *pipelines.Config, dryRunEnabled bool, noPush bool) error {
+func imagePublishPipeline(stdout io.Writer, stderr io.Writer, config *pipelines.Config, dryRunEnabled bool) error {
 	pipeline := pipelines.NewimagePublish(stdout, stderr)
 	pipeline.DryRunEnabled = dryRunEnabled
-	pipeline.NoPush = noPush
 
 	return pipeline.WithConfig(config).Run()
 }
@@ -192,8 +204,16 @@ func codeScanPipeline(stdout io.Writer, stderr io.Writer, config *pipelines.Conf
 	return pipeline.WithConfig(config).Run()
 }
 
+func deployPipeline(stdout io.Writer, stderr io.Writer, config *pipelines.Config, dryRunEnabled bool) error {
+	pipeline := pipelines.NewDeploy(stdout, stderr)
+	pipeline.DryRunEnabled = dryRunEnabled
+
+	return pipeline.WithConfig(config).Run()
+}
+
 func debugPipeline(stdout io.Writer, stderr io.Writer, dryRunEnabled bool) error {
 	pipeline := pipelines.NewDebug(stdout, stderr)
 	pipeline.DryRunEnabled = dryRunEnabled
+
 	return pipeline.Run()
 }