+++ pkg/pipelines/code-scan.go
@@ -19,6 +19,14 @@ type CodeScan struct {
 	SemgrepErrorOnFindingsEnabled bool
 	SemgrepRules                  string
 	config                        *Config
+	runtime                       struct {
+		gitleaksFile     *os.File
+		semgrepFile      *os.File
+		bundleFilename   string
+		gitleaksFilename string
+		semgrepFilename  string
+		gatecheckListBuf *bytes.Buffer
+	}
 }
 
 func (s *CodeScan) WithConfig(config *Config) *CodeScan {
@@ -35,63 +43,102 @@ func NewCodeScan(stdout io.Writer, stderr io.Writer) *CodeScan {
 	}
 }
 
-func (p *CodeScan) Run() error {
-	var gitleaksError, semgrepError error
-
-	gitleaksFilename := path.Join(p.config.ArtifactsDir, p.config.CodeScan.GitleaksFilename)
-	semgrepFilename := path.Join(p.config.ArtifactsDir, p.config.CodeScan.SemgrepFilename)
-	slog.Info("run image scan pipeline", "dry_run_enabled", p.DryRunEnabled, "artifact_directory", p.config.ArtifactsDir)
+// preRun is responsible for opening file needed during the run
+func (p *CodeScan) preRun() error {
+	var err error
 
 	if err := MakeDirectoryP(p.config.ArtifactsDir); err != nil {
 		slog.Error("failed to create artifact directory", "name", p.config.ArtifactsDir)
 		return errors.New("Code Scan Pipeline failed to run. See log for details.")
 	}
 
-	slog.Debug("open gitleaks artifact for output", "filename", gitleaksFilename)
-
-	gitleaksFile, err := OpenOrCreateFile(gitleaksFilename)
+	p.runtime.gitleaksFilename = path.Join(p.config.ArtifactsDir, p.config.CodeScan.GitleaksFilename)
+	p.runtime.gitleaksFile, err = OpenOrCreateFile(p.runtime.gitleaksFilename)
 	if err != nil {
-		slog.Error("cannot open gitleaks report file", "filename", gitleaksFilename, "error", err)
-		return errors.New("Code Scan Pipeline failed, Gitleaks did not run. See log for details.")
+		slog.Error("cannot open gitleaks report file", "filename", p.runtime.gitleaksFile, "error", err)
+		return err
 	}
-	defer gitleaksFile.Close()
+
+	p.runtime.semgrepFilename = path.Join(p.config.ArtifactsDir, p.config.CodeScan.SemgrepFilename)
+	p.runtime.semgrepFile, err = OpenOrCreateFile(p.runtime.semgrepFilename)
+	if err != nil {
+		slog.Error("cannot open semgrep report file", "filename", p.runtime.semgrepFilename, "error", err)
+		return err
+	}
+
+	if err := InitGatecheckBundle(p.config, p.Stderr, p.DryRunEnabled); err != nil {
+		slog.Error("cannot initialize gatecheck bundle", "error", err)
+		return err
+	}
+
+	p.runtime.gatecheckListBuf = new(bytes.Buffer)
+	p.runtime.bundleFilename = path.Join(p.config.ArtifactsDir, p.config.GatecheckBundleFilename)
+
+	return nil
+}
+
+func (p *CodeScan) Run() error {
+	if !p.config.CodeScan.Enabled {
+		slog.Warn("code scan pipeline is disabled, skip.")
+		return nil
+	}
+
+	if err := p.preRun(); err != nil {
+		return errors.New("Code Scan Pipeline Pre-Run Failed. See log for details.")
+	}
+
+	defer func() {
+		_ = p.runtime.gitleaksFile.Close()
+		_ = p.runtime.semgrepFile.Close()
+	}()
+
+	slog.Info("run image scan pipeline", "dry_run_enabled", p.DryRunEnabled, "artifact_directory", p.config.ArtifactsDir)
+
+	slog.Debug("open gatecheck bundle file for output", "filename", p.runtime.bundleFilename)
 
 	// All of the gatcheck summaries should print at the end
-	summaryBuf := new(bytes.Buffer)
 	buf := new(bytes.Buffer)
 	// MultiWriter will write to the gitleaks file and to the buf so gatecheck can parse it
-	mw := io.MultiWriter(gitleaksFile, buf)
-	gitleaksError = RunGitleaksDetect(mw, p.Stderr, p.config, p.DryRunEnabled)
+	mw := io.MultiWriter(p.runtime.gitleaksFile, buf)
+	gitleaksError := RunGitleaksDetect(mw, p.Stderr, p.config, p.DryRunEnabled)
 
 	slog.Debug("summarize gitleaks report")
-	err = RunGatecheckList(summaryBuf, buf, p.Stderr, "gitleaks", p.DryRunEnabled)
-	if err != nil {
+	if err := RunGatecheckList(p.runtime.gatecheckListBuf, buf, p.Stderr, "gitleaks", p.DryRunEnabled); err != nil {
 		slog.Error("cannot run gatecheck list on gitleaks report")
 	}
-	// Add a new line to separate the reports
-	fmt.Fprintln(summaryBuf, "")
 
-	semgrepFile, err := OpenOrCreateFile(semgrepFilename)
-	if err != nil {
-		slog.Error("failed to create semgrep scan report file")
-		return errors.New("Code Scan Pipeline failed, Semgrep did not run. See log for details.")
-	}
-	defer semgrepFile.Close()
+	// Add a new line to separate the reports
+	fmt.Fprintln(p.runtime.gatecheckListBuf, "")
 
 	buf = new(bytes.Buffer)
-	mw = io.MultiWriter(semgrepFile, buf)
+	mw = io.MultiWriter(p.runtime.semgrepFile, buf)
 
-	semgrepError = RunSemgrep(mw, p.Stderr, p.config, p.DryRunEnabled, p.SemgrepExperimental)
+	semgrepError := RunSemgrep(mw, p.Stderr, p.config, p.DryRunEnabled, p.SemgrepExperimental)
 	slog.Debug("summarize semgrep report")
-	err = RunGatecheckList(summaryBuf, buf, p.Stderr, "semgrep", p.DryRunEnabled)
-	if err != nil {
+	if err := RunGatecheckList(p.runtime.gatecheckListBuf, buf, p.Stderr, "semgrep", p.DryRunEnabled); err != nil {
 		slog.Error("cannot run gatecheck list on semgrep report")
 	}
 
-	// print the summaries
-	_, _ = summaryBuf.WriteTo(p.Stdout)
+	var postRunError error
 
-	return errors.Join(gitleaksError, semgrepError)
+	if err := p.postRun(); err != nil {
+		postRunError = errors.New("Code Scan Pipeline Post-Run Failed. See log for details.")
+	}
+
+	return errors.Join(gitleaksError, semgrepError, postRunError)
+}
+
+func (p *CodeScan) postRun() error {
+	files := []string{p.runtime.gitleaksFilename, p.runtime.semgrepFilename}
+	err := RunGatecheckBundleAdd(p.runtime.bundleFilename, p.Stderr, p.DryRunEnabled, files...)
+	if err != nil {
+		slog.Error("cannot run gatecheck bundle add", "error", err)
+	}
+
+	// print the Gatecheck List Content
+	_, _ = p.runtime.gatecheckListBuf.WriteTo(p.Stdout)
+
+	return err
 }
 
 // RunGitleaksDetect the report will be written to stdout
@@ -105,8 +152,12 @@ func RunGitleaksDetect(reportDst io.Writer, stdErr io.Writer, config *Config, dr
 	if err != nil {
 		return err
 	}
-
 	tempReportFilename := reportFile.Name()
+
+	defer func() {
+		_ = reportFile.Close()
+		_ = os.Remove(tempReportFilename)
+	}()
 
 	cmd := shell.GitleaksCommand(nil, nil, stdErr).DetectSecrets(config.CodeScan.GitleaksSrcDir, tempReportFilename)
 	err = cmd.WithDryRun(dryRunEnabled).Run()